+++ packages/sui-svg/templates/icon-component.js
@@ -2,7 +2,10 @@ const template = (code, config, state) => {
   return `import React, {memo} from 'react'
 import AtomIcon from '@s-ui/react-atom-icon'
 
-export default memo(props => <AtomIcon {...props}>${code}</AtomIcon>)
+const MemoAtomIcon = memo(props => <AtomIcon {...props}>${code}</AtomIcon>)
+MemoAtomIcon.displayName = 'AtomIcon'
+
+export default MemoAtomIcon
 `
 }